+++ fixtures/gists-app/tests/server-html-test.ts
@@ -1,7 +1,7 @@
 import type { Browser, Page } from "puppeteer";
 import puppeteer from "puppeteer";
 
-import { prettyHtml } from "./utils";
+import { getHtml } from "./utils";
 
 const testPort = 3000;
 const testServer = `http://localhost:${testPort}`;
@@ -16,12 +16,14 @@ describe("the server HTML", () => {
 
   afterEach(() => browser.close());
 
-  it("is correct", async () => {
-    await page.goto(`${testServer}/`);
+  describe("for the root URL", () => {
+    it("is correct", async () => {
+      await page.goto(`${testServer}/`);
 
-    // Important: Do NOT wait for React to hydrate
-    // because we want to test the server HTML output.
+      // Important: Do NOT wait for React to hydrate
+      // because we want to test the server HTML output.
 
-    expect(prettyHtml(await page.content())).toMatchSnapshot("page");
+      expect(await getHtml(page)).toMatchSnapshot();
+    });
   });
 });