+++ fixtures/gists-app/tests/utils.ts
@@ -1,5 +1,21 @@
 import prettier from "prettier";
 import type { Page, Response } from "puppeteer";
+import cheerio from "cheerio";
+
+export async function getHtml(page: Page, selector?: string): Promise<string> {
+  let html = await page.content();
+  return prettyHtml(selector ? selectHtml(html, selector) : html);
+}
+
+export function selectHtml(source: string, selector: string): string {
+  let el = cheerio(selector, source);
+
+  if (!el.length) {
+    throw new Error(`No element matches selector "${selector}"`);
+  }
+
+  return cheerio.html(el);
+}
 
 export function prettyHtml(source: string): string {
   return prettier.format(source, { parser: "html" });